requirements.txt
psycopg2-binarydjango remove
psycopg2-binarydjango remove